% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_nas}
\alias{add_nas}
\title{A function that adds NA values to a vector or pair of vectors}
\usage{
add_nas(x, y = NULL, nas_indices_removed)
}
\arguments{
\item{x}{Vector to add NA's to}

\item{y}{Optional second vector to add NA's to}

\item{nas_indices_removed}{Indices where NA's had previously been removed
(AKA indices where NA's will be in the
*output* vector)}
}
\value{
A list containing: x with NA's added

                           y with NA's added
}
\description{
NA's will be added to x and y at indices specified where they had 
been previously removed
}
