% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{canbe.numeric}
\alias{canbe.numeric}
\title{A better version of is.numeric}
\usage{
canbe.numeric(x)
}
\arguments{
\item{x}{Vector to check}
}
\value{
TRUE if \code{x} is numeric or can be converted to numeric without
        warnings. Otherwise, FALSE
}
\description{
This function works like is.numeric but also checks if x can be coerced to 
numeric without warnings. If is.numeric or can be coerced to numeric, 
returns TRUE. Otherwise, returns
}
