% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{checkdim_inputs}
\alias{checkdim_inputs}
\title{Check dimension of inputs}
\usage{
checkdim_inputs(
  input,
  input_name,
  needed_len,
  needed_name = "the number of files"
)
}
\arguments{
\item{input}{Values to be used}

\item{input_name}{Name of the input used for error messages}

\item{needed_len}{Desired length of output vector}

\item{needed_name}{What the desired length corresponds to (e.g. number of files)}
}
\value{
The values of \code{input} coerced to a vector of length \code{needed_len}
}
\description{
Check if values are a vector of given length, if not coerce to be so
Returns error messages using appropriate names
}
