% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dots_parser}
\alias{dots_parser}
\title{A function that parses dots arguments and only passes the correct ones}
\usage{
dots_parser(FUN, ...)
}
\arguments{
\item{FUN}{The function to be called}

\item{...}{Additional arguments, some of which may not be arguments
for \code{FUN}}
}
\value{
The output of \code{FUN} operating on arguments in \code{...}
}
\description{
Use this in cases where a parent function calls multiple sub-functions
and passes the ... dots argument to each, but some arguments users
specify in the ... dots argument only work for some of the sub-functions.
In this case, dots_parser will check and run \code{FUN} with only
the arguments that \code{FUN} accepts
}
