% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_design_tidy}
\alias{example_design_tidy}
\title{Design for example growth curve data
 
A tidy-shaped dataset with the experimental design (i.e. plate layout)
for the example data included with \code{gcplyr}.}
\format{
A dataframe with 96 rows and 3 variables:
\describe{
     \item{Well}{The well of the plate}
     \item{Bacteria_strain}{The numbered bacterial strain growing in each well}
     \item{Phage}{Whether or not the bacteria were simulated growing with phages}
 }
}
\usage{
example_design_tidy
}
\description{
Wells A1...A8 through F1...F8 contain 48 different
simulated bacterial strains growing alone. Wells G1...G8 through L1...L8
contain the same 48 bacterial strains in an identical layout, but this
time growing in the presence of a phage
}
\keyword{datasets}
