% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_widedata}
\alias{example_widedata}
\title{Example growth curve data in wide format}
\format{
A dataframe with ___ rows and 97 variables:
\describe{
     \item{time}{time, in seconds, since growth curve began}
     \item{A1, A2...H11, H12}{bacterial density in the given well}
 }
}
\usage{
example_widedata
}
\description{
A dataset containing example growth of 96 wells of simulated bacteria
 or bacteria and phages
}
\keyword{datasets}
