% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fill_data_metadata}
\alias{fill_data_metadata}
\title{Fill output data.frame with \code{data} and \code{metadata}}
\usage{
fill_data_metadata(
  output,
  input,
  rs,
  metadata_include = 1:length(input$metadata)
)
}
\arguments{
\item{output}{Data frame for data and metadata to be filled into}

\item{input}{One data and metadata to fill into \code{output}}

\item{rs}{index of the row in \code{output} where data and metadata should
start being filled in}

\item{metadata_include}{Indices of which metadata items should be included

Use NULL for no metadata inclusion}
}
\value{
A list, where [[1]] is \code{output} containing the data and
        metadata from \code{input} but properly formatted for writing.
        
        And [[2]] is the updated \code{rs} to be used for subsequent
        calls to \code{fill_data_metadata}
}
\description{
This is an internal function used by \code{write_blocks} that does the 
fill data and metadata step
}
