% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_next_extrema}
\alias{find_next_extrema}
\title{Find the next extrema from the current position (for internal use only)}
\usage{
find_next_extrema(
  cnt_pos,
  y,
  width_limit = NULL,
  height_limit = NULL,
  looking_for = c("minima", "maxima")
)
}
\arguments{
\item{cnt_pos}{Current position index to start search from}

\item{y}{Numeric vector of y values in which to identify local extrema}

\item{width_limit}{Width of the window (in number of \code{y} values) used to
search for local extrema. A narrower width will me more
sensitive to narrow local maxima/minima, while a wider
width will be less sensitive to local maxima/minima.}

\item{height_limit}{The maximum change in \code{y} a single extrema-search
step is allowed to take.
For example, a maxima-finding function will not pass a
valley deeper than height_limit. This also limits
approaches to true extrema, so if it is set too small
the function may return non-extrema}

\item{looking_for}{Is the next window looking for a "maxima" or a "minima"?}
}
\value{
The index of the next extrema, of the same type as specified by
        looking_for
}
\description{
Find the next extrema from the current position (for internal use only)
}
