% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{first_below}
\alias{first_below}
\title{Find the first point when a numeric vector falls below some threshold}
\usage{
first_below(y, x = NULL, threshold, return = "index", subset = NULL)
}
\arguments{
\item{y}{Numeric vector of y values in which to identify first point
below some threshold}

\item{x}{Optional numeric vector of corresponding x values}

\item{threshold}{Threshold y value of interest}

\item{return}{One of c("index", "x"), determining whether the function
will return the index or x value associated with the
first y value below some threshold}

\item{subset}{A vector of Boolean values indicating which x and y values
should be included (TRUE) or excluded (FALSE).

If returning index, index will be for the whole vector and
not the subset of the vector}
}
\description{
This function takes a vector of \code{x} and \code{y} values and 
returns the index (by default) of the first point that falls
below some threshold y value.
}
\details{
This function is designed to be compatible with dplyr::group_by and summarize


This function is designed to be compatible with dplyr::group_by and summarize
}
