% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_window_limits}
\alias{get_window_limits}
\title{Find limits of window for next extrema search (for internal use only)}
\usage{
get_window_limits(cnt_pos, y, width_limit = NULL, height_limit = NULL)
}
\arguments{
\item{cnt_pos}{Current position index to start search from}

\item{y}{Numeric vector of y values in which to identify local extrema
(technically optional if only using width_limit, but since
find_next_extrema will always provide it's required for simplicity)}

\item{width_limit}{Width of the window (in number of \code{y} values) used to
search for local extrema. A narrower width will me more
sensitive to narrow local maxima/minima, while a wider
width will be less sensitive to local maxima/minima.}

\item{height_limit}{The maximum change in \code{y} a single extrema-search
step is allowed to take.
For example, a maxima-finding function will not pass a
valley deeper than height_limit. This also limits
approaches to true extrema, so if it is set too small
the function may return non-extrema}
}
\value{
Vector, of length two, with start index and end index of the
        next window to use
}
\description{
Find limits of window for next extrema search (for internal use only)
}
