% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_windows}
\alias{get_windows}
\title{A function that gets windows for moving-window like calculations}
\usage{
get_windows(
  x,
  y,
  window_width_n = NULL,
  window_width = NULL,
  window_height = NULL,
  edge_NA,
  force_height_multi_n = FALSE
)
}
\arguments{
\item{x}{Vector of x values}

\item{y}{Vector of y values}

\item{window_width_n}{Width of the window (in number of \code{y} values).}

\item{window_width}{Width of the window (in units of \code{x}).}

\item{window_height}{The maximum change in \code{y} within each window.}

\item{edge_NA}{Boolean for whether windows that pass the edge of the data
should be returned as NA or simply truncated at the edge
of the data.

For instance, moving average-like functions typically want
edge_NA = TRUE so as not to calculate a moving-average
on points near the edge of the domain with smaller n

In contrast, extrema-finding typically want edge_NA = FALSE
so that searching simply stops at the edge of the data}

\item{force_height_multi_n}{Boolean for whether windows limits set by \code{window_height}
should always have at least 3 data points in them (or 2 
data points for windows located at the edge of the domain)
         
This is necessary for window-finding with window_height 
because otherwise you can end up with windows that contain 
only a single data-point, counting as both a max and a min}
}
\value{
A list of vectors, where each vector contains the indices
        of the data in the window centered at that point
}
\description{
A function that gets windows for moving-window like calculations
}
