% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_blockdesign}
\alias{import_blockdesign}
\title{Import blockdesigns from file (work-in-progress)}
\usage{
import_blockdesign(
  files,
  startrow = 2,
  startcol = 2,
  metadata = list(design_element = c(1, 1)),
  field_sep = "_",
  id_cols = c("block_name", "design_element"),
  ...
)
}
\arguments{
\item{files}{tbd}

\item{startrow}{tbd}

\item{startcol}{tbd}

\item{metadata}{tbd}

\item{field_sep}{tbd}

\item{id_cols}{tbd}

\item{...}{tbd}
}
\description{
This function imports blockdesigns from file into the R environment
}
