% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{infer_block_metadata}
\alias{infer_block_metadata}
\title{A function that infers whether blocks have nested metadata}
\usage{
infer_block_metadata(blocks)
}
\arguments{
\item{blocks}{The list of blocks}
}
\value{
TRUE or FALSE whether the blocks contain metadata
        (also passes warning when inferring, or error when unable to)
}
\description{
A function that infers whether blocks have nested metadata
}
