% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lag_time}
\alias{lag_time}
\title{calculate lag time}
\usage{
lag_time(rate, y0, x1, y1, trans_y = "log")
}
\arguments{
\item{rate}{Numeric value or vector of per-capita growth rate (slope)
to use when calculating lag time}

\item{y0}{Numeric value or vector of y value (density) to find the 
intersection of the growth rate line}

\item{x1}{x value (time) when rate was achieved}

\item{y1}{y value (density) when rate was achieved}

\item{trans_y}{One of \code{c("linear", "log")} specifying the
                transformation of y-values.

                \code{'log'} is the default, producing calculations of
                lag time assuming a transition to exponential growth
                
                \code{'linear'} is available for alternate uses}
}
\value{
A scalar or vector of the lag time(s)
}
\description{
This function takes a vector of \code{x} and \code{y} values
and returns a scalar for the area under the curve, calculated using 
the trapezoid rule
}
\details{
This function is designed to be compatible for use within
 dplyr::group_by and dplyr::summarize
}
