% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{moving_average}
\alias{moving_average}
\title{Moving average smoothing}
\usage{
moving_average(formula, data, window_width_n)
}
\arguments{
\item{formula}{Formula specifying the numeric response (density) 
and numeric predictor (time).}

\item{data}{Dataframe containing variables in \code{formula}}

\item{window_width_n}{Number of data points wide the moving average window is
(therefore, must be an odd number of points)}
}
\value{
Vector of smoothed data, with NA's appended at both ends
}
\description{
This function uses a moving average to smooth data
}
