% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print_df}
\alias{print_df}
\title{Nicely print the contents of a data.frame}
\usage{
print_df(x, col.names = FALSE, row.names = FALSE)
}
\arguments{
\item{x}{The \code{data.frame} to be printed}

\item{col.names}{Boolean for whether column names should be printed}

\item{row.names}{Boolean for whether row names should be printed}
}
\description{
This function uses \code{write.table} to print the input \code{data.frame}
in a nicely-formatted manner that is easy to read
}
