% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_blocks}
\alias{read_blocks}
\title{Read blockmeasures}
\usage{
read_blocks(
  files,
  extension = NULL,
  startrow = NULL,
  endrow = NULL,
  startcol = NULL,
  endcol = NULL,
  sheet = NULL,
  metadata = NULL,
  block_names = NULL,
  header = NA,
  sider = NA,
  wellnames_Excel = TRUE
)
}
\arguments{
\item{files}{A vector of filepaths relative to the current working directory
where each filepath is a single plate read}

\item{extension}{(optional) the extension of the files:
"csv", "xls", or "xlsx"
If none provided, \code{read_blocks} will infer file extension from
provided filenames}

\item{startrow, endrow, startcol, endcol}{(optional) the rows and columns where 
the measures data are in \code{files},
can be a vector or list the same length as \code{files}, or
a single value that applies to all \code{files}.
If not provided data is presumed to begin on the first
row and column of the files.}

\item{sheet}{(optional) If data is in .xls or .xlsx files, which sheet it 
is located on. Defaults to the first sheet if not specified}

\item{metadata}{(optional) non-spectrophotometric data that should be associated
with each read blockmeasures. A list where each item in the
list is a vector of length 2. Each vector should provide the
row and column where the metadata is located in the blockmeasures
input file. If the list is named those names will be inherited
to the output metadata.}

\item{block_names}{(optional) vector of names corresponding to each plate
in \code{files}. If not provided, block_names are inferred
from the filenames}

\item{header}{\code{TRUE}, \code{FALSE}, or \code{NA}, or a vector of
such values, indicating whether the file(s) contains the
column names as its first line. If \code{header = NA}
will attempt to infer the presence of column names. If
\code{header = FALSE} or no column names are inferred when 
\code{header = NA}, column names will be generated
automatically according to \code{wellnames_Excel}}

\item{sider}{\code{TRUE}, \code{FALSE}, or \code{NA}, or a vector of
such values, indicating whether the file(s) contains the
row names as its first line. If \code{sider = NA}
will attempt to infer the presence of row names. If
\code{sider = FALSE} or no row names are inferred when 
\code{sider = NA}, row names will be generated
automatically according to \code{wellnames_Excel}}

\item{wellnames_Excel}{If row names and column names are not provided in the
input dataframe as specified by \code{header}
and \code{sider}, then names will be generated
automatically.
If \code{wellnames_Excel} is TRUE, generated names
will use Excel-style base-26 lettering for columns
and numbers for rows. 
If \code{wellnames_Excel} is FALSE, rows and columns
will be numbered with "R" and "C" prefixes, respectively.}
}
\value{
A list where each entry is a list containing the block measures data
        followed by the block_names (or filenames, if block_names is not 
        provided) and any specified metadata.
}
\description{
A function that reads block measures into the R environment
}
