% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reorder}
\alias{reorder}
\title{A function that reorders x and y based on x}
\usage{
reorder(x = NULL, y)
}
\arguments{
\item{x}{Vector to reorder based on}

\item{y}{Vector to reorder based on x}
}
\value{
A list containing: the reordered x

                           the reordered y
                           
                           the original order, such that 
                           return[["x"]][order(return[["order"]])] == x and
                           return[["y"]][order(return[["order"]])] == y
}
\description{
A function that reorders x and y based on x
}
