% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reorder_xy}
\alias{reorder_xy}
\title{A function that reorders x and y based on x}
\usage{
reorder_xy(x = NULL, y)
}
\arguments{
\item{x}{Vector to reorder based on}

\item{y}{Vector to reorder based on x}
}
\value{
A list containing: 

    [1] "x" = the reordered x

    [2] "y" = the reordered y
                           
    [3] "order" = the original order, such that:
                           
    \code{return[["x"]][order(return[["order"]])] == x}
                           
    and
                           
    \code{return[["y"]][order(return[["order"]])] == y}
}
\description{
A function that reorders x and y based on x
}
