% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rm_nas}
\alias{rm_nas}
\title{A function that removes na's values from a vector or pair of vectors}
\usage{
rm_nas(x = NULL, y = NULL, na.rm, stopifNA = FALSE)
}
\arguments{
\item{x, y}{Vectors to remove NA's from. At least one must be non-NULL}

\item{na.rm}{Boolean, should NA's be removed}

\item{stopifNA}{Boolean, should an error be passed if na.rm = FALSE
and there are NA's in x or y?}
}
\value{
A list containing: x with NA's removed

                           y with NA's removed
                         
                           vector of indices where NA's were removed
                           (NULL when none were removed)
}
\description{
Any values of x and y will be removed for indices where x or y are NA
}
