% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sep_finder}
\alias{sep_finder}
\title{A function that finds a character not present in block data nor metadata}
\usage{
sep_finder(blocks, nested_metadata = NULL)
}
\arguments{
\item{blocks}{The list of blocks}

\item{nested_metadata}{A Boolean for if there is nested metadata in
\code{blocks}. Will attempt to infer if left NULL}
}
\value{
vector of characters not found in the blocks that can be used
        as a separator without issue (or error if none can be found)
}
\description{
Typically this is used to find a separator that can be used without
accidentally implying separations in the data that shouldn't be there
}
