% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trans_block_to_wide}
\alias{trans_block_to_wide}
\title{Transform blocks to wides}
\usage{
trans_block_to_wide(
  blocks,
  wellnames_sep = "",
  nested_metadata = NULL,
  colnames_first = FALSE
)
}
\arguments{
\item{blocks}{Blocks, either a single data.frame or a list of
data.frames}

\item{wellnames_sep}{String to use as separator for well names between 
rowname and column name (ordered according to
\code{colnames_first}}

\item{nested_metadata}{A logical indicating the existence of nested metadata
in the \code{blockmeasures} list, e.g. as is typically
output by \code{read_blocks}. If NULL, will attempt to
infer existence of nested metadata}

\item{colnames_first}{In the wellnames created by \code{paste}-ing the
rownames and column names, should the column names
come first}
}
\value{
A single widemeasures data.frame
}
\description{
Takes blocks and returns them in a wide format
}
