% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trans_wide_to_block}
\alias{trans_wide_to_block}
\title{Transform wides into block format (in progress)}
\usage{
trans_wide_to_block(wides, collapse = NULL, wellnames_sep = "_")
}
\arguments{
\item{wides}{Wide dataframe(s) (either widemeasures or widedesign)}

\item{collapse}{NULL or a string to use for concatenating contents
together. If NULL each row in \code{wides} will be put into 
its own block. If a string, that string will be used to 
\code{paste} together all elements and all elements will
be returned in a single block}

\item{wellnames_sep}{A string used to identify the rownames and column
names}
}
\value{
A list of block data.frames (if \code{collapse} is not 
        \code{NULL} the list is of length 1)
}
\description{
Takes wides and returns them in a block format
}
