% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_blocks}
\alias{write_blocks}
\title{Write block designs to csv}
\usage{
write_blocks(blocks, file, ...)
}
\arguments{
\item{file}{The filename (if a single design), or vector of file names
(if multiple designs)}

\item{...}{Other arguments passed to \code{write.csv}}

\item{designs}{data.frame or matrix (if one design), or list of data.frames
or matrices, to be written to file}
}
\value{
Nothing, but R objects are written to files
}
\description{
This function is basically just a wrapper for write.csv that works
 for block-shaped data.frames
}
\details{
If \code{designs} is a list, \code{names(designs)} will be placed in the
1,1 cell of each block
}
