% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_deriv}
\alias{calc_deriv}
\title{Calculate derivatives of vector of data}
\usage{
calc_deriv(y, x = NULL, x_scale = 1, percapita = FALSE, subset_by = NULL)
}
\arguments{
\item{y}{Data to calculate difference or derivative of}

\item{x}{Vector of x values provided as a simple numeric
(e.g. if time, in number of seconds, not POSIX)}

\item{x_scale}{Factor to scale x by in denominator of derivative calculation

When \code{x_scale = NA}, \code{calc_deriv} returns 
differences in \code{y}

When \code{x_scale = 1}, \code{calc_deriv} returns the 
standard derivative in the same units as y/x

For other units, set x_scale to the ratio of the the units of 
x to the desired units. E.g. if x is in seconds, but the 
desired derivative is in units of /minute, set 
\code{x_scale = 60} (since there are 60 seconds in 1 minute).}

\item{percapita}{When percapita = TRUE, the differences of y are 
divided by y}

\item{subset_by}{if subset_by is provided, it should be a vector (same 
length as y), the unique values of which will 
separate calculations}
}
\value{
A vector of the difference, derivative, or per-capita derivative 
        values the same length as \code{y}, with \code{NA} appended to 
        the end
}
\description{
Provided a vector of y values, this function returns the difference
between sequential values, derivative between sequential values,
or per-capita derivative between sequential values
}
