% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_local_extrema}
\alias{find_local_extrema}
\title{Find local extrema of numeric vector}
\usage{
find_local_extrema(
  y,
  x = NULL,
  return = "index",
  return_maxima = TRUE,
  return_minima = TRUE,
  return_endpoints = TRUE,
  width_limit = NULL,
  width_limit_n = NULL,
  height_limit = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{y}{Numeric vector of y values in which to identify local extrema}

\item{x}{Optional numeric vector of corresponding x values}

\item{return}{One of c("index", "x", "y"), determining whether the function
will return the index, x value, or y value associated with the
identified extremas}

\item{return_maxima, return_minima}{Boolean for which classes of local extrema
to return}

\item{return_endpoints}{Should the first and last values in \code{y}
be included if they are in the returned 
vector of extrema?}

\item{width_limit}{Width of the window (in units of \code{x}) used to
search for local extrema. A narrower width will be more
sensitive to narrow local maxima/minima, while a wider
width will be less sensitive to local maxima/minima.}

\item{width_limit_n}{The maximum number of data points a single 
extrema-search step is allowed to take. For example,
when maxima-finding, the function will not pass
a valley consisting of more than \code{width_limit_n}
data points.

A smaller \code{width_limit_n} will be more sensitive 
to narrow local maxima/minima, while a larger 
\code{width_limit_n} will be less sensitive to 
narrow local maxima/minima.}

\item{height_limit}{The maximum change in \code{y} a single extrema-search
step is allowed to take.  For example, when 
maxima-finding, the function will not pass a
valley deeper than \code{height_limit}.

A smaller \code{height_limit} will be more sensitive 
to shallow local maxima/minima, while a larger 
\code{height_limit} will be less sensitive to 
shallow maxima/minima.}

\item{na.rm}{Boolean whether NA's should be removed before analyzing}
}
\value{
If \code{return = "index"}, a vector of indices corresponding 
          to local extrema in the data
          
        If \code{return = "x"}, a vector of x values corresponding
          to local extrema in the data
         
        If \code{return = "y"}, a vector of y values corresponding
          to local extrema in the data
}
\description{
This function takes a vector of \code{y} values and returns a vector
of the indices of all local value extrema (by default, this includes both
local minima and local maxima)
Either width_limit_n or height_limit must be provided
}
\details{
If multiple of width_limit, width_limit_n, and height_limit are provided, 
steps are limited conservatively (a single step must meet all criteria)

This function is designed to be compatible for use within
 dplyr::group_by and dplyr::summarize
}
