% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_local_extrema}
\alias{find_local_extrema}
\title{Find local extrema of numeric vector}
\usage{
find_local_extrema(
  y,
  return_maxima = TRUE,
  return_minima = TRUE,
  width_limit = NULL,
  height_limit = NULL,
  remove_endpoints = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{y}{Numeric vector of y values in which to identify local extrema}

\item{return_maxima, return_minima}{Boolean for which classes of local extrema
to return}

\item{width_limit}{Width of the window (in number of \code{y} values) used to
search for local extrema. A narrower width will me more
sensitive to narrow local maxima/minima, while a wider
width will be less sensitive to local maxima/minima.}

\item{height_limit}{The maximum change in \code{y} a single extrema-search
step is allowed to take.
For example, a maxima-finding function will not pass a
valley deeper than height_limit. This also limits
approaches to true extrema, so if it is set too small
the function may return non-extrema}

\item{remove_endpoints}{Should the first and last values in \code{y}
be removed from the returned vector of extrema}

\item{na.rm}{Boolean whether NA's should be removed before analyzing}
}
\value{
A vector of indices of \code{y} corresponding to local extrema
        in the data
}
\description{
This function takes a vector of \code{y} values and returns a vector
of the indices of all local value extrema (by default, this includes both
local minima and local maxima)
Either width_limit or height_limit must be provided
}
\details{
If both width_limit and height_limit are provided, steps are limited
conservatively (a single step must meet both criteria)

This function is designed to be compatible with dplyr::group_by and summarize
}
