% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{first_peak}
\alias{first_peak}
\title{Find the first local peak of a numeric vector}
\usage{
first_peak(y, x = NULL, return = "index", width_limit = 0.2 * length(values))
}
\arguments{
\item{y}{Numeric vector of y values in which to identify local extrema}

\item{x}{Optional numeric vector of corresponding x values}

\item{return}{One of c("index", "x", "y"), determining whether the function
will return the index, x value, or y value associated with the
first peak in y values}

\item{width_limit}{Width of the window (in number of \code{y}) used to
search for local extrema. A narrower width will me more
sensitive to local maxima/minima, while a wider
width will be less sensitive to local maxima/minima.}

\item{...}{Other parameters to pass to \code{find_local_extrema}}
}
\description{
This function takes a vector of \code{y} values and returns the index
(by default) of the first local maxima. It serves as a wrapper function
for \code{find_local_extrema}
}
\details{
This function is designed to be compatible with dplyr::group_by and summarize
}
