% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_blockdesign}
\alias{import_blockdesign}
\title{Import blockdesigns}
\usage{
import_blockdesign(files, ...)
}
\arguments{
\item{files}{Vector of filenames (as strings), each of which is a 
block-shaped designs file. Inputs can be .csv, .xls, or .xlsx}

\item{...}{Other arguments to pass to \code{read_blocks}, 
\code{paste_blocks}, \code{trans_block_to_wide},
\code{trans_wide_to_tidy}, or \code{separate_tidy}.

See Details for more information}
}
\description{
Function to import block-shaped designs from files and return tidy designs.
This function acts as a wrapper to call read_blocks, paste_blocks, 
trans_block_to_wide, and trans_wide_to_tidy in one go
}
\details{
Common arguments that you may want to provide include:

             startrow, endrow, startcol, endcol, sheet - specifying the
             location of design information inside \code{files} to 
             \code{read_blocks}
             
             sep - string separating separate design elements for
             \code{separate_tidys} in block design files that are already
             pasted
             
             into - vector of column names for design elements after
             separation for \code{separate_tidys}
}
