% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_blockdesigns}
\alias{import_blockdesigns}
\title{Import blockdesigns}
\usage{
import_blockdesigns(files, into = NULL, ...)
}
\arguments{
\item{files}{Vector of filenames (as strings), each of which is a 
block-shaped designs file. Inputs can be .csv, .xls, or .xlsx}

\item{into}{Vector of column names for design elements after
separation by \code{separate_tidys}. Should be in the same
order as elements listed within \code{files} and/or same order
as corresponding \code{files} themselves. If \code{NULL},
column names will simply be "Design_1", "Design_2", etc.}

\item{...}{Other arguments to pass to \code{read_blocks}, 
\code{paste_blocks}, \code{trans_block_to_wide},
\code{trans_wide_to_tidy}, or \code{separate_tidy}.

See Details for more information}
}
\description{
Function to import block-shaped designs from files and return tidy designs.
This function acts as a wrapper to call \code{read_blocks}, 
\code{paste_blocks}, \code{trans_block_to_wide}, \code{trans_wide_to_tidy}, 
and \code{separate_tidys} in one go
}
\details{
Common arguments that you may want to provide include:

             startrow, endrow, startcol, endcol, sheet - specifying the
             location of design information inside \code{files} to 
             \code{read_blocks}
             
             sep - string separating separate design elements for
             \code{separate_tidys} in block design files that are already
             pasted
}
