% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_blockmeasures}
\alias{import_blockmeasures}
\title{Import blockmeasures}
\usage{
import_blockmeasures(
  files,
  num_plates = 1,
  plate_names = NULL,
  wellnames_sep = "_",
  ...
)
}
\arguments{
\item{files}{Vector of filenames (as strings), each of which is a blockmeasures
formatted file. Inputs can be .csv, .xls, or .xlsx}

\item{num_plates}{Number of plates. If multiple plates uninterleave will be
used to separate blockmeasures into those plates accordingly}

\item{plate_names}{(optional) Names to put onto the plates when output}

\item{wellnames_sep}{String to use as separator for well names between 
rowname and column name}

\item{...}{Other arguments to pass to \code{read_blocks}, \code{uninterleave},
or \code{widen_blocks}}
}
\description{
Function to import blockmeasures from files and return widemeasures
This function acts as a wrapper to call read_blocks, uninterleave, 
then widen_blocks in one go
}
