% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{infer_names}
\alias{infer_names}
\title{An internal function that handles name inference logic for other functions}
\usage{
infer_names(df, startrow, endrow, startcol, endcol, header, sider)
}
\arguments{
\item{df}{The dataframe}

\item{startrow, endrow, startcol, endcol}{The rows & columns specified by user}

\item{header, sider}{Whether the file contains a header or sider
(can be TRUE, FALSE, or NA)}
}
\value{
a list: list(startrow, endrow, startcol, endcol, rownames_col, colnames_row)
}
\description{
This function takes in the fully-read dataframe alongside information
about whether rows and columns have been specified, and whether header
and sider have been specified.
}
\details{
None of the specified arguments should be a vector, they should
         all be single values or NA's
}
