% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{moving_average}
\alias{moving_average}
\title{Moving average smoothing}
\usage{
moving_average(
  formula,
  data,
  window_width_n = NULL,
  window_width = NULL,
  window_width_frac = NULL,
  na.rm = TRUE,
  warn_nonnumeric_sort = TRUE
)
}
\arguments{
\item{formula}{Formula specifying the numeric response (density) 
and numeric predictor (time).}

\item{data}{Dataframe containing variables in \code{formula}}

\item{window_width_n}{Number of data points wide the moving average window is
(therefore, must be an odd number of points)}

\item{window_width}{Width of the moving average window (in units of \code{x})}

\item{window_width_frac}{Width of the window (as a fraction of the total
number of data points).}

\item{na.rm}{logical whether NA's should be removed before analyzing}

\item{warn_nonnumeric_sort}{logical whether warning should be issued when 
predictor variable that data is sorted by is 
non-numeric.}
}
\value{
Vector of smoothed data, with NA's appended at both ends
}
\description{
This function uses a moving average to smooth data
}
