% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{moving_median}
\alias{moving_median}
\title{Moving median smoothing}
\usage{
moving_median(formula, data, window_width)
}
\arguments{
\item{formula}{Formula specifying the numeric response (density) 
and numeric predictor (time).}

\item{data}{Dataframe containing variables in \code{formula}}

\item{window_width}{Number of data points wide the moving average window is
(therefore, must be an odd number of points)}
}
\value{
Vector of smoothed data, with NA's appended at both ends
}
\description{
This function uses a moving median to smooth data
}
